+++ httpie/__init__.py
@@ -3,7 +3,7 @@ HTTPie: modern, user-friendly command-line HTTP client for the API era.
 
 """
 
-__version__ = '4.0.0.b1'
-__date__ = '2024-01-01'
+__version__ = '4.0.0'
+__date__ = '2024-06-25'
 __author__ = 'Jakub Roztocil'
 __licence__ = 'BSD'